Dust removal. Simplified time handling code, other fixes.
*Add locking
*Use preferred values
*Improvements in events and latency handling
*Various minor fixes such as style and code readability.

+++ lib/ClientNode.h
@@ -138,7 +138,7 @@ private:
 				bigtime_t	 fLastLateness;
 
 				bigtime_t	 fTime;
-				size_t		 fFramesSent;
+				int64		 fFramesSent;
 
 				BBufferGroup*	fBufferGroup;
 };
